backend/requirements.txt
Flask==2.0.1
requests==2.25.1
psycopg2-binary==2.8.6
werkzeug == 2.0.1
flask_cors == 3.0.10
bcrypt == 2.0.


